rf: delete duplicate tests, move some non-duplicate tests to the relevant class, correcte the expected result in several tests
