+++ src/test/java/school/redrover/StartPageTest.java
@@ -4,13 +4,10 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 import school.redrover.page.home.HomePage;
 import school.redrover.runner.BaseTest;
-import school.redrover.runner.TestUtils;
 
 import java.util.List;
 
 public class StartPageTest extends BaseTest {
-
-    private static final String NEW_FOLDER_NAME = "FirstFolder";
 
     @Test
     public void testStartPageMainPanelContent() {
@@ -42,43 +39,6 @@ public class StartPageTest extends BaseTest {
                 startPageSideContent,
                 List.of("http://localhost:8080/view/all/newJob", "http://localhost:8080/view/all/builds",
                         "http://localhost:8080/manage", "http://localhost:8080/me/my-views"));
-    }
-
-    @Test
-    public void testCreateNewFolder() {
-        TestUtils.createFolder(getDriver(), NEW_FOLDER_NAME);
-
-        List<String> projectList = new HomePage(getDriver())
-                .getItemList();
-
-        Assert.assertListContainsObject(
-                projectList,
-                NEW_FOLDER_NAME,
-                "Folder is not created");
-    }
-
-    @Test(dependsOnMethods = "testCreateNewFolder")
-    public void testDeleteNewFolder() {
-
-        String welcomeText = new HomePage(getDriver())
-                .openFolder(NEW_FOLDER_NAME)
-                .clickDeleteButtonSidebarAndConfirm()
-                .getWelcomeDescriptionText();
-
-        Assert.assertEquals(welcomeText,
-                "This page is where your Jenkins jobs will be displayed. To get started, you can set up distributed builds or start building a software project.");
-    }
-
-    @Test
-    public void testDeleteNewFolderViaChevron() {
-        TestUtils.createFolder(getDriver(), NEW_FOLDER_NAME);
-
-        String welcomeText = new HomePage(getDriver())
-                .selectDeleteFromItemMenuAndClickYes(NEW_FOLDER_NAME)
-                .getWelcomeDescriptionText();
-
-        Assert.assertEquals(welcomeText,
-                "This page is where your Jenkins jobs will be displayed. To get started, you can set up distributed builds or start building a software project.");
     }
 
     @Test
